python/requirements.txt
mysql-connector-python
paramiko
cryptography